06: Both parts
The solution is identical between them.

The subtle hints about exponential growth and "index 0" in the puzzle
indicated that just simulating every fish wasn't going to be feasible.

Instead, keep a circular buffer of 9 days, and just store the number of
fish for each day in each slot in the buffer.