08: Part 2 quadruple nested for loop?!
I'm not sure whether it would have been better to convert the strings
to bit-masks and count bits instead of letters.

I'm not sure if there's a better approach than these deeply nested
loops.

I feel like having the letters in different orders in the left and right
sides of '|' is a bit mean, but perhaps that's meant to hint to use
bits instead of letters?