18: Part 1. This kicked my butt!
I started doing a tree, but then figured that was probably a trap
because of the left/right traversal.

It took me ages to settle on tokenisation, and then I failed to parse
the rule about doing *all* explosions before *any* splits, and then
failed to implement that when I had realised.